pyproject.toml
[project]
name = "stock"
version = "0.1.0"
description = "NVIDIA Stock Checker using official API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
]
